separate db models
