+++ src/app/database/models/users.py
@@ -1,0 +1,21 @@
+from sqlalchemy import Boolean, Column, ForeignKey, Integer, String
+
+from src.app.database.base import BaseModel
+
+
+class User(BaseModel):
+    __tablename__ = "users"
+
+    id = Column(Integer, primary_key=True, index=True)
+    username = Column(String, unique=True, index=True)
+    hashed_password = Column(String)
+    is_active = Column(Boolean, default=True)
+
+
+class UserBooks(BaseModel):
+    __tablename__ = "user_books"
+
+    id = Column(Integer, primary_key=True, index=True)
+    user_id = Column(Integer, ForeignKey("users.id"))
+    book_id = Column(Integer, ForeignKey("books.id"))
+    count = Column(Integer)
